pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pollen-forecast"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Antoine TAVANT", email = "antoine.tavant@ecoco2.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "cdsapi",
  "pandas",
  "xarray",
  "netcdf4",
  "matplotlib",
  "watchfiles",
  # "cartopy",
  "setuptools",
  "django>=4.2.20",
  "altair>=5.4.1",
  "geoip2>=4.8.1",
  "django-q2>=1.7.6",
  "blessed>=1.20.0",
  "pytest-django>=4.10.0",
  "pytest-sugar>=1.0.0",
  "djangorestframework>=3.15.2",
  "markdown>=3.7",
  "django-filter>=24.3",
  "psycopg2-binary>=2.9.10",
]

[project.urls]
Documentation = "https://github.com/Antoine TAVANT/pollen-forecast#readme"
Issues = "https://github.com/Antoine TAVANT/pollen-forecast/issues"
Source = "https://github.com/Antoine TAVANT/pollen-forecast"

[tool.hatch.version]
path = "src/pollen_forecast/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/pollen_forecast tests}"

[tool.coverage.run]
source_pkgs = ["pollen_forecast", "tests"]
branch = true
parallel = true
omit = [
  "src/pollen_forecast/__about__.py",
]

[tool.coverage.paths]
pollen_forecast = ["src/pollen_forecast", "*/pollen-forecast/src/pollen_forecast"]
tests = ["tests", "*/pollen-forecast/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.hatch.envs.prod]
dependencies = [
  "setuptools",
]
[tool.hatch.envs.prod.scripts]
serve = "panel serve src/pollen_forecast/panel_app.py --address 0.0.0.0 --port 5006 --allow-websocket-origin=antoinetavant.fr"

[tool.hatch.envs.taipy]
dependencies = [
  "taipy",
  "plotly"
]

[tool.hatch.envs.taipy.scripts]
serve = "python src/pollen_forecast/taipy_app.py"


[tool.hatch.envs.streamlit]
dependencies = [
  "streamlit",
  "streamlit-searchbox",
  "streamlit-geolocation",
]

[tool.hatch.envs.streamlit.scripts]
serve = "streamlit run --server.port 5006 --server.baseUrlPath pollen  --browser.gatherUsageStats false --server.runOnSave true   src/pollen_forecast/Météo_pollen.py"

[tool.pytest.ini_options]
log_cli = true
addopts = "--nomigrations --reuse-db -vv"
pythonpath = "src/pollen_forecast/djangoserver"
DJANGO_SETTINGS_MODULE = "meteopollen.settings"
# -- recommended but optional:
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
